Fixed issue where all servers' logs would go in the same file
Removed log4cpp.properties and changed all source references to use
log4cpp::PatternLayout and log4cpp::FileAppender.

Fixed bug in config.json.in where there was no comma inbetween the
server and options objects.

Added a default server location so if the user does not give an absolute
path for the server path, it assumes that the server name given is a
location the default server folder.